requirements.txt
Django==1.4
django-piston==0.2.2
python-dateutil==1.5
-e hg+https://bitbucket.org/ubernostrum/django-registration/#egg=django-registration
django-debug-toolbar
